rag-service/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.0
weaviate-client==4.4.1
openai==1.12.0
sentence-transformers==2.3.1
pydantic==2.6.0
python-dotenv==1.0.1
httpx==0.26.0
